chore: consolidate live test harnesses into tests/live/ + reconcile tool count (BAT-1144 Part 1)
Part 1 of the BAT-1144 test-infra work. Consolidates the scattered live/manual
test folders into a single coherent tests/live/ root. tests/nodejs-project/
(CI-gated unit tests) is unchanged.

  testing/                   -> tests/live/anthropic/   (Anthropic live probes + lib + FINDINGS)
  testing/xai-oauth-spike/   -> tests/live/oauth-spikes/
  tests/xai-models/          -> tests/live/xai/         (live-models.test.js -> live-models.probe.js)
  tests/live/openai/_shared/ -> tests/live/_shared/     (shared env/fixture for all probes)

- Deleted testing/test-telegram-draft.js (misfiled OpenClaw port).
- Rewrote import depths in moved files: anthropic bundle imports ../app -> ../../../app,
  openai _shared ./ -> ../. lib.js loadEnv is __dirname-relative, so credential loading
  is unaffected by the move.
- Swept every stale path ref across the repo (production comments in ai.js/database.js/
  claude.js/xai.js, CLAUDE.md, DIAGNOSTICS.md, the BAT-1124 contract, 4 nodejs-project
  tests). Zero stale refs remain. Production-file changes are comment/string only -- no
  logic changed.

Tool-count reconcile (the harness README flagged this as a follow-up): authoritative
count is 64 built-in tools (telegram channel, no MCP), from the assembled
require('tools/index').TOOLS.length -- asserted live on every harness --self-check run.
The old "66"/"67" were miscounts: raw `grep -c input_schema tools/*.js` returns 67,
inflated by 3 comment mentions of "input_schema" (agent_pay.js x2, system.js x1) on top
of the 64 real schema keys. Fixed PROJECT.md (63 -> 64) + the harness note.

Also adds the CodeRabbit badge to the README (per request).

Test plan:
- node tests/nodejs-project/smoke.js -- PASS (27/27 modules)
- node tests/live/openai/live-models.probe.js --self-check -- PASS (apikey + oauth)
- 4 edited CI tests (claude-adaptive-thinking, claude-thinking-signature,
  recent-sessions-heartbeat-filter, xai) -- ALL PASS
- import-resolution check: all rewritten relative paths resolve to real files
