+++ app/src/main/assets/nodejs-project/DIAGNOSTICS.md
@@ -240,7 +240,7 @@ grep -i "OAuth refresh\|oauth_refresh\|invalid_grant" node_debug.log | tail -20
 ### xAI Grok — grok-4.5 slow / "not responding" (reasoning bound + first-touch provisioning)
 **Symptoms:** grok-4.5 hangs then fails while grok-4.3 replies fine. `node_debug.log` shows `[Trace] … timeoutSource:"transport" status:-1 error:"Timeout"` (~60s per attempt) on a grok-4.5 turn; OR a brand-new model 403s on the first message then works minutes later.
 **Diagnosis — two distinct, known causes:**
-- **Unbounded reasoning (fixed BAT-1124):** xAI honors the OpenAI-style `reasoning_effort` **string**, NOT OpenRouter's `reasoning:{effort}` **object**. With neither, grok-4.5 reasons UNBOUNDED — on a big agent request (many tools + system prompt) that is >60s of *silent* reasoning, tripping the 60s socket-idle timeout; grok-4.3 (lighter) answers in ~3s. `providers/xai.js formatRequest` now always sends `reasoning_effort` (`minimal` for heartbeats, `high` when the user enables reasoning, else `low`). Reproduce/verify with `node tests/xai-models/live-models.test.js --diagnose`.
+- **Unbounded reasoning (fixed BAT-1124):** xAI honors the OpenAI-style `reasoning_effort` **string**, NOT OpenRouter's `reasoning:{effort}` **object**. With neither, grok-4.5 reasons UNBOUNDED — on a big agent request (many tools + system prompt) that is >60s of *silent* reasoning, tripping the 60s socket-idle timeout; grok-4.3 (lighter) answers in ~3s. `providers/xai.js formatRequest` now always sends `reasoning_effort` (`minimal` for heartbeats, `high` when the user enables reasoning, else `low`). Reproduce/verify with `node tests/live/xai/live-models.probe.js --diagnose`.
 - **First-touch provisioning:** a brand-new xAI model 403s on its very FIRST request (API access provisions lazily), then works a few minutes later — the same 403→200 pattern a fresh account shows on grok-4.3. The 403 retry-once grace can be too short for this; if a 403 persists on a model you know exists, wait ~1 min and retry before assuming a permanent tier-gate.
 **Fix/check:** confirm `formatRequest` emits `reasoning_effort`; for a first-touch 403 on a known-good model, retry after a minute.
