+++ docs/internal/contracts/BAT-1124-xai-grok-oauth-contract.md
@@ -34,7 +34,7 @@ Reusing the Grok-CLI public client_id `b1a00492-073a-47ea-816f-4c329264a828` is 
 - xAI enforces entitlement server-side (the 403 tier-gate), so a subscription lacking API access fails gracefully to the API-key path — same fallback story as any provider. Optional/non-blocking: ask xAI dev-support for explicit third-party sanction.
 
 ### D4 — Model list — **concrete IDs are a pre-device-test gate (blocker), registry stays static**
-The registry is **static JSON bundled in the APK** (no runtime `/v1/models` discovery). With `freeform:false`, `models:[]` is **not shippable** (unlike `openrouter`, which is `freeform:true`). At impl time, enumerate `GET https://api.x.ai/v1/models` with a live token, seed `model-registry.json`, and set a valid `defaultModel` **in the model list**. Also add the `xai` branch to the hardcoded `config.js:320-323` default-model chain — its `else` returns `claude-opus-4-8`, which `api.x.ai` would **404**. Confirm whether oauth vs api_key expose different model sets → add `modelsByAuth` if so. **CONFIRMED via live SuperGrok `/v1/models` (2026-07-07, 9 models):** text/chat = `grok-4.3` (**default** — tested 200), `grok-4.20-0309-reasoning`, `grok-4.20-0309-non-reasoning`, `grok-4.20-multi-agent-0309`, `grok-build-0.1`; image/video (**OUT** of v1 text scope) = `grok-imagine-image`, `grok-imagine-image-quality`, `grok-imagine-video`, `grok-imagine-video-1.5`. Registry seeds the **5 text models, default `grok-4.3`**. (This is the OAuth token's view; api-key parity unconfirmed — add `modelsByAuth` only if a console key differs.) **Post-GATE-0 update (2026-07-09) — registry trimmed to `grok-4.3` + `grok-4.5`:** per product call the registry ships ONLY these two (anything else = user-typed Custom model); the `grok-4.20-*` + `grok-build-0.1` entries were dropped, and `grok-4.20-multi-agent-0309` was in fact **broken on chat/completions** anyway (live test: *"Multi Agent requests are not allowed on chat completions"*). **grok-4.5 device-test found two failures, both root-caused via `tests/xai-models` (live OAuth harness) + `node_debug.log`:** (1) **first-touch 403** — a brand-new model provisions lazily and 403s the first message, then works (same 403→200 pattern grok-4.3 showed on a fresh account); (2) **unbounded reasoning** — the app sent NO reasoning param, and xAI honors the OpenAI-style **`reasoning_effort` string** (NOT OpenRouter's `reasoning:{effort}` object), so grok-4.5 reasoned silently >60s on the big agent request → the 60s socket-idle timeout fired (grok-4.3 answered in ~3s). **Fix:** `providers/xai.js formatRequest` now always sends `reasoning_effort` (`minimal` heartbeats / `high` user-enabled / else `low`); pinned by `xai.test.js` + reproduced by `--diagnose`. **Default stays `grok-4.3` until the grok-4.5 reasoning fix is device-verified**, then flip to grok-4.5. Mobile context cap 200000 for all.
+The registry is **static JSON bundled in the APK** (no runtime `/v1/models` discovery). With `freeform:false`, `models:[]` is **not shippable** (unlike `openrouter`, which is `freeform:true`). At impl time, enumerate `GET https://api.x.ai/v1/models` with a live token, seed `model-registry.json`, and set a valid `defaultModel` **in the model list**. Also add the `xai` branch to the hardcoded `config.js:320-323` default-model chain — its `else` returns `claude-opus-4-8`, which `api.x.ai` would **404**. Confirm whether oauth vs api_key expose different model sets → add `modelsByAuth` if so. **CONFIRMED via live SuperGrok `/v1/models` (2026-07-07, 9 models):** text/chat = `grok-4.3` (**default** — tested 200), `grok-4.20-0309-reasoning`, `grok-4.20-0309-non-reasoning`, `grok-4.20-multi-agent-0309`, `grok-build-0.1`; image/video (**OUT** of v1 text scope) = `grok-imagine-image`, `grok-imagine-image-quality`, `grok-imagine-video`, `grok-imagine-video-1.5`. Registry seeds the **5 text models, default `grok-4.3`**. (This is the OAuth token's view; api-key parity unconfirmed — add `modelsByAuth` only if a console key differs.) **Post-GATE-0 update (2026-07-09) — registry trimmed to `grok-4.3` + `grok-4.5`:** per product call the registry ships ONLY these two (anything else = user-typed Custom model); the `grok-4.20-*` + `grok-build-0.1` entries were dropped, and `grok-4.20-multi-agent-0309` was in fact **broken on chat/completions** anyway (live test: *"Multi Agent requests are not allowed on chat completions"*). **grok-4.5 device-test found two failures, both root-caused via `tests/live/xai` (live OAuth harness) + `node_debug.log`:** (1) **first-touch 403** — a brand-new model provisions lazily and 403s the first message, then works (same 403→200 pattern grok-4.3 showed on a fresh account); (2) **unbounded reasoning** — the app sent NO reasoning param, and xAI honors the OpenAI-style **`reasoning_effort` string** (NOT OpenRouter's `reasoning:{effort}` object), so grok-4.5 reasoned silently >60s on the big agent request → the 60s socket-idle timeout fired (grok-4.3 answered in ~3s). **Fix:** `providers/xai.js formatRequest` now always sends `reasoning_effort` (`minimal` heartbeats / `high` user-enabled / else `low`); pinned by `xai.test.js` + reproduced by `--diagnose`. **Default stays `grok-4.3` until the grok-4.5 reasoning fix is device-verified**, then flip to grok-4.5. Mobile context cap 200000 for all.
 
 ### D5 — Flavor parity with OpenAI OAuth (both flavors)
 Ship the `oauth` path on **both flavors** (`dappStore` + `googlePlay`), exactly as the OpenAI OAuth ships today (no `BuildConfig.DISTRIBUTION` gate) — for consistency with our existing treatment. Payments policy isn't implicated (externally-purchased subscription, same shape as the shipped ChatGPT OAuth). If xAI ever objects, we can flavor-gate later; no gating for v1.
