+++ tests/live/anthropic/test-auth.js
