+++ tests/live/anthropic/test-headers.js
