+++ tests/live/anthropic/test-production-shape.js
